Remove old browser workarounds, split convert.js
